Reviewed-on: https://git.sh-edraft.de/sh-edraft.de/sh_cpl/pulls/67

+++ src/cpl_cli/command/add_service.py
@@ -22,8 +22,8 @@ class AddService(CommandABC):
         CommandABC.__init__(self)
 
         self._config = config
-
         self._workspace = workspace
+        self._is_simulation = False
 
     @property
     def help_message(self) -> str:
@@ -36,8 +36,9 @@ class AddService(CommandABC):
             target-project:  Name of the project to be referenced
         """)
 
-    @staticmethod
-    def _edit_project_file(source: str, project_settings: ProjectSettings, build_settings: BuildSettings):
+    def _edit_project_file(self, source: str, project_settings: ProjectSettings, build_settings: BuildSettings):
+        if self._is_simulation:
+            return
         with open(source, 'w') as file:
             file.write(json.dumps({
                 ProjectSettings.__name__: SettingsHelper.get_project_settings_dict(project_settings),
@@ -51,6 +52,11 @@ class AddService(CommandABC):
         :param args:
         :return:
         """
+        if 'simulate' in args:
+            args.remove('simulate')
+            Console.write_line('Running in simulation mode:')
+            self._is_simulation = True
+
         if len(args) == 0:
             Console.error('Expected source and target project')
             return
@@ -60,7 +66,7 @@ class AddService(CommandABC):
             return
 
         elif len(args) > 2:
-            Console.error(f'Unexpected argument: {" ".join(args[2:])}')
+            Console.error(f'Unexpected argument(s): {", ".join(args[2:])}')
             return
 
         # file names
@@ -104,14 +110,14 @@ class AddService(CommandABC):
             Console.error(f'Invalid target: {target}')
             return
 
-        if target in build_settings.project_references:
-            Console.error(f'Project reference already exists.')
-            return
-
         if self._workspace is None:
             target = f'../{target}'
         else:
             target = target.replace('src', '..')
+
+        if target in build_settings.project_references:
+            Console.error(f'Project reference already exists.')
+            return
 
         build_settings.project_references.append(target)
